BugFix: Fixed recursion errors (both searching same target twice, and searching a second user before first has resolved
