+++ src/App.js
@@ -1,11 +1,17 @@
+import { useState } from "react";
 import "./App.css";
 import SearchFunction from "./components/SearchFunction";
+import UserContext, { defaultUserContextData } from "./context/UserContext";
 
 function App() {
+    const [userData, setUserData] = useState(defaultUserContextData)
+    
     return (
         <div className="App">
             <h1>League of Legends Challenge Info</h1>
-            <SearchFunction />
+            <UserContext.Provider  value={{userData, setUserData}}>
+                <SearchFunction />
+            </UserContext.Provider>
         </div>
     );
 }
